moved ProjectWithComplexAnalyzersTests to its own class
